+++ ava.config.js
@@ -1,0 +1,11 @@
+export default {
+  extensions: {
+    ts: "module"
+  },
+  nodeArguments: [
+    "--loader=ts-node/esm/transpile-only"
+  ],
+  files: [
+    "src/**/*.test.ts"
+  ]
+}
