--- lib/index.ts
@@ -1,7 +1,0 @@
-export * from "./FormDataEncoder"
-
-// Utilities
-export * from "./FileLike"
-export * from "./FormDataLike"
-export * from "./util/isFileLike"
-export * from "./util/isFormData"
