+++ package.json
@@ -1,4 +1,5 @@
 {
+  "type": "module",
   "name": "form-data-encoder",
   "version": "1.7.2",
   "description": "Encode FormData content into the multipart/form-data format",
@@ -13,22 +14,17 @@
     "spec-compatible",
     "form"
   ],
-  "main": "./lib/cjs/index.js",
-  "module": "./lib/esm/index.js",
+  "main": "./lib/index.js",
+  "module": "./lib/index.js",
+  "exports": "./lib/esm/index.js",
   "types": "./@type/index.d.ts",
-  "exports": {
-    "import": "./lib/esm/index.js",
-    "require": "./lib/cjs/index.js"
-  },
   "scripts": {
     "eslint": "eslint lib/**/*.ts",
     "staged": "lint-staged",
     "coverage": "c8 npm test",
     "ci": "c8 npm test && c8 report --reporter=json",
-    "build:esm": "ttsc --project tsconfig.esm.json",
-    "build:cjs": "ttsc --project tsconfig.cjs.json",
-    "build:types": "ttsc --project tsconfig.d.ts.json",
-    "build": "npm run build:esm && npm run build:cjs && npm run build:types",
+    "build:types": "tsc --project tsconfig.d.ts.json",
+    "build": "tsc && npm run build:types",
     "test": "ava --fail-fast",
     "cleanup": "npx rimraf @type \"lib/**/*.js\"",
     "prepare": "npm run cleanup && npm run build",
@@ -42,16 +38,16 @@
     "@octetstream/eslint-config": "5.0.0",
     "@types/mime-types": "2.1.1",
     "@types/node": "17.0.25",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
-    "@zoltu/typescript-transformer-append-js-extension": "1.0.1",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "ava": "4.2.0",
     "c8": "7.11.2",
-    "eslint": "7.32.0",
+    "eslint": "8.13.0",
     "eslint-config-airbnb-typescript": "12.3.1",
-    "eslint-plugin-ava": "12.0.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.29.4",
+    "eslint-import-resolver-typescript": "2.7.1",
+    "eslint-plugin-ava": "13.2.0",
+    "eslint-plugin-jsx-a11y": "6.4.1",
+    "eslint-plugin-react": "7.26.1",
     "formdata-node": "4.3.2",
     "husky": "7.0.4",
     "lint-staged": "12.4.0",
