+++ src/__helper__/readStream.ts
