+++ src/util/escapeName.test.ts
@@ -1,31 +1,31 @@
 import test from "ava"
 
-import escapeName from "./escapeName"
+import escapeName from "./escapeName.js"
 
 const CR = "%0D"
 const LF = "%0A"
 const Q = "%22"
 
 test("Escapes all the CRs in the name", t => {
-  t.is<string>(escapeName("\rna\rme\r"), `${CR}na${CR}me${CR}`)
+  t.is<string, string>(escapeName("\rna\rme\r"), `${CR}na${CR}me${CR}`)
 })
 
 test("Keeps escaped CR as is", t => {
   const expected = `name${CR}`
 
-  t.is<string>(escapeName(expected), expected)
+  t.is<string, string>(escapeName(expected), expected)
 })
 
 test("Escapes all the LFs in the name", t => {
-  t.is<string>(escapeName("nam\ne\n"), `nam${LF}e${LF}`)
+  t.is<string, string>(escapeName("nam\ne\n"), `nam${LF}e${LF}`)
 })
 
 test("Keeps escaped LF as is", t => {
   const expected = `name${LF}`
 
-  t.is<string>(escapeName(expected), expected)
+  t.is<string, string>(escapeName(expected), expected)
 })
 
 test("Escapes all double quotes in the name", t => {
-  t.is<string>(escapeName("\"name\""), `${Q}name${Q}`)
+  t.is<string, string>(escapeName("\"name\""), `${Q}name${Q}`)
 })
