+++ src/util/isFunction.ts
