+++ src/util/isPlainObject.ts
