+++ tsconfig.ava.json
@@ -1,7 +1,6 @@
 {
-  "extends": "./tsconfig.cjs.json",
+  "extends": "./tsconfig.json",
   "ts-node": {
-    "compiler": "ttypescript",
     "transpileOnly": true,
     "preferTsExts": true
   }
