+++ tsconfig.d.ts.json
@@ -1,5 +1,5 @@
 {
-  "extends": "./tsconfig.esm.json",
+  "extends": "./tsconfig.json",
   "compilerOptions": {
     "declaration": true,
     "declarationDir": "./@type",
